refactor(monitor): deprecate Monitor in favor of Monitor1D
